+++ TennisApp/src/main/java/cz/inqool/tennisapp/domain/reservation/ReservationResponse.java
@@ -7,14 +7,14 @@ import java.time.LocalDateTime;
 
 @Data
 public class ReservationResponse {
-    private Long id;
+    private int id;
     private String courtName;
     private String customerName;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
 
     public ReservationResponse(Reservation reservation) {
-        this.id = (long) reservation.getId();
+        this.id = reservation.getId();
         this.courtName = reservation.getCourt().getName();
         this.customerName = reservation.getCustomer().getName();
         this.startTime = reservation.getStartTime();
